move out null/memory/random
