+++ Cargo.toml
@@ -52,9 +52,12 @@ members = [
     "src/query/storages/hive-meta-store",
     "src/query/storages/index",
     "src/query/storages/information-schema",
+    "src/query/storages/memory",
     "src/query/storages/preludes",
     "src/query/storages/factory",
     "src/query/storages/fuse-result",
+    "src/query/storages/null",
+    "src/query/storages/random",
     "src/query/storages/share",
     "src/query/storages/view",
     "src/query/streams",
