+++ src/query/service/Cargo.toml
@@ -66,7 +66,10 @@ common-storages-fuse-result = { path = "../storages/fuse-result" }
 common-storages-hive = { path = "../storages/hive", optional = true }
 common-storages-index = { path = "../storages/index" }
 common-storages-information-schema = { path = "../storages/information-schema" }
+common-storages-memory = { path = "../storages/memory" }
+common-storages-null = { path = "../storages/null" }
 common-storages-preludes = { path = "../storages/preludes" }
+common-storages-random = { path = "../storages/random" }
 common-storages-share = { path = "../storages/share" }
 common-storages-view = { path = "../storages/view" }
 common-streams = { path = "../streams" }
