+++ src/query/service/tests/it/storages/memory.rs
@@ -23,9 +23,9 @@ use common_planner::extras::Extras;
 use common_planner::extras::Statistics;
 use common_planner::plans::Projection;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_storages_memory::MemoryTable;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::create_table_v2::TableOptions;
-use databend_query::storages::memory::MemoryTable;
 use databend_query::stream::DataBlockStream;
 use futures::TryStreamExt;
