+++ src/query/storages/memory/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "common-storages-preludes"
-version = "0.1.0"
-edition = "2021"
-authors = ["Databend Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
+name = "common-storages-memory"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
@@ -28,11 +28,13 @@ common-pipeline-sources = { path = "../../pipeline/sources" }
 common-pipeline-transforms = { path = "../../pipeline/transforms" }
 common-planner = { path = "../../planner" }
 common-storage = { path = "../../../common/storage" }
+common-storages-view = { path = "../view" }
 common-users = { path = "../../users" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
 backon = "0.2"
 chrono = "0.4.22"
+futures = "0.3.24"
 itertools = "0.10.5"
 once_cell = "1.15.0"
 opendal = { version = "0.19", features = ["layers-retry", "layers-tracing", "layers-metrics", "compress"] }
