+++ src/query/storages/memory/src/memory_table.rs
@@ -17,6 +17,9 @@ use std::collections::HashSet;
 use std::collections::VecDeque;
 use std::sync::Arc;
 
+use common_catalog::catalog::StorageDescription;
+use common_catalog::table::Table;
+use common_catalog::table_context::TableContext;
 use common_datablocks::DataBlock;
 use common_datablocks::InMemoryData;
 use common_datavalues::ColumnRef;
@@ -27,6 +30,12 @@ use common_datavalues::TypeID;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_app::schema::TableInfo;
+use common_pipeline_core::processors::port::OutputPort;
+use common_pipeline_core::processors::processor::ProcessorPtr;
+use common_pipeline_core::Pipeline;
+use common_pipeline_sinks::processors::sinks::ContextSink;
+use common_pipeline_sources::processors::sources::SyncSource;
+use common_pipeline_sources::processors::sources::SyncSourcer;
 use common_planner::extras::Extras;
 use common_planner::extras::Statistics;
 use common_planner::plans::Projection;
@@ -37,16 +46,7 @@ use once_cell::sync::Lazy;
 use parking_lot::Mutex;
 use parking_lot::RwLock;
 
-use crate::pipelines::processors::port::OutputPort;
-use crate::pipelines::processors::processor::ProcessorPtr;
-use crate::pipelines::processors::ContextSink;
-use crate::pipelines::processors::SyncSource;
-use crate::pipelines::processors::SyncSourcer;
-use crate::pipelines::Pipeline;
-use crate::sessions::TableContext;
-use crate::storages::memory::memory_part::MemoryPartInfo;
-use crate::storages::StorageDescription;
-use crate::storages::Table;
+use crate::memory_part::MemoryPartInfo;
 
 static IN_MEMORY_DATA: Lazy<Arc<RwLock<InMemoryData<u64>>>> =
     Lazy::new(|| Arc::new(Default::default()));
